+++ aiida/backends/tests/__init__.py
@@ -7,6 +7,7 @@
 # For further information on the license, see the LICENSE.txt file        #
 # For further information please visit http://www.aiida.net               #
 ###########################################################################
+# pylint: disable=line-too-long
 """
 Module for defining tests that required access to (a temporary) database
 """
@@ -105,12 +106,6 @@ DB_TEST_LIST = {
         'engine.utils': ['aiida.backends.tests.engine.test_utils'],
         'engine.work_chain': ['aiida.backends.tests.engine.test_work_chain'],
         'engine.workfunctions': ['aiida.backends.tests.engine.test_workfunctions'],
-        'export_migration.migration': ['aiida.backends.tests.tools.importexport.migration.test_migration'],
-        'export_migration.v01_to_v02': ['aiida.backends.tests.tools.importexport.migration.test_v01_to_v02'],
-        'export_migration.v02_to_v03': ['aiida.backends.tests.tools.importexport.migration.test_v02_to_v03'],
-        'export_migration.v03_to_v04': ['aiida.backends.tests.tools.importexport.migration.test_v03_to_v04'],
-        'export_migration.v04_to_v05': ['aiida.backends.tests.tools.importexport.migration.test_v04_to_v05'],
-        'export_migration.v05_to_v06': ['aiida.backends.tests.tools.importexport.migration.test_v05_to_v06'],
         'generic': ['aiida.backends.tests.test_generic'],
         'manage.configuration.config.': ['aiida.backends.tests.manage.configuration.test_config'],
         'manage.configuration.migrations.': ['aiida.backends.tests.manage.configuration.migrations.test_migrations'],
@@ -151,6 +146,13 @@ DB_TEST_LIST = {
         'tools.importexport.prov_redesign': ['aiida.backends.tests.tools.importexport.test_prov_redesign'],
         'tools.importexport.simple': ['aiida.backends.tests.tools.importexport.test_simple'],
         'tools.importexport.specific_import': ['aiida.backends.tests.tools.importexport.test_specific_import'],
+        'tools.importexport.migration.migration': ['aiida.backends.tests.tools.importexport.migration.test_migration'],
+        'tools.importexport.migration.v01_to_v02': ['aiida.backends.tests.tools.importexport.migration.test_v01_to_v02'],
+        'tools.importexport.migration.v02_to_v03': ['aiida.backends.tests.tools.importexport.migration.test_v02_to_v03'],
+        'tools.importexport.migration.v03_to_v04': ['aiida.backends.tests.tools.importexport.migration.test_v03_to_v04'],
+        'tools.importexport.migration.v04_to_v05': ['aiida.backends.tests.tools.importexport.migration.test_v04_to_v05'],
+        'tools.importexport.migration.v05_to_v06': ['aiida.backends.tests.tools.importexport.migration.test_v05_to_v06'],
+        'tools.importexport.orm.attributes': ['aiida.backends.tests.tools.importexport.orm.test_attributes'],
         'tools.importexport.orm.calculations': ['aiida.backends.tests.tools.importexport.orm.test_calculations'],
         'tools.importexport.orm.codes': ['aiida.backends.tests.tools.importexport.orm.test_codes'],
         'tools.importexport.orm.comments': ['aiida.backends.tests.tools.importexport.orm.test_comments'],
