+++ aiida/backends/tests/tools/importexport/migration/test_v03_to_v04.py
@@ -19,12 +19,12 @@ import zipfile
 
 from aiida.backends.testbase import AiidaTestCase
 from aiida.backends.tests.utils.archives import get_archive_file, get_json_files
-from aiida.cmdline.utils.migration.utils import verify_metadata_version
-from aiida.cmdline.utils.migration.v03_to_v04 import migrate_v3_to_v4
 from aiida.common.archive import extract_tar, extract_zip
 from aiida.common.exceptions import NotExistent
 from aiida.common.folders import SandboxFolder
 from aiida.common.json import load as jsonload
+from aiida.tools.importexport.migration.utils import verify_metadata_version
+from aiida.tools.importexport.migration.v03_to_v04 import migrate_v3_to_v4
 
 
 class TestMigrateV03toV04(AiidaTestCase):
