+++ aiida/backends/tests/tools/importexport/migration/test_v04_to_v05.py
@@ -19,12 +19,12 @@ import zipfile
 
 from aiida.backends.testbase import AiidaTestCase
 from aiida.backends.tests.utils.archives import get_archive_file, get_json_files
-from aiida.cmdline.utils.migration.utils import verify_metadata_version
-from aiida.cmdline.utils.migration.v04_to_v05 import migrate_v4_to_v5
 from aiida.common.archive import extract_tar, extract_zip
 from aiida.common.exceptions import NotExistent
 from aiida.common.folders import SandboxFolder
 from aiida.common.json import load as jsonload
+from aiida.tools.importexport.migration.utils import verify_metadata_version
+from aiida.tools.importexport.migration.v04_to_v05 import migrate_v4_to_v5
 
 
 class TestMigrateV04toV05(AiidaTestCase):
