+++ aiida/tools/importexport/migration/v04_to_v05.py
@@ -28,7 +28,7 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import absolute_import
 
-from aiida.cmdline.utils.migration.utils import verify_metadata_version, update_metadata, remove_fields
+from aiida.tools.importexport.migration.utils import verify_metadata_version, update_metadata, remove_fields
 
 
 def migration_drop_node_columns_nodeversion_public(metadata, data):
